+++ NEWS.md
@@ -1,5 +1,9 @@
 # User visible changes in `OptimPackNextGen` package
 
+## Version 0.4.2
+
+- Fix loading of `Zygote` extension with Julia ≤ 1.8.
+
 ## Version 0.4.1
 
 - In Powell's methods (COBYLA, NEWUOA, and BOBYQA):